Update jackett-update.sh
Rewritten to handle download failures and retry 5 times if needed. It'll email an address of your choice if it fails.

+++ mediacentre/Configs/jackett-update.sh
@@ -1,15 +1,37 @@
 #!/bin/bash
-systemctl stop jackett
+# Destination email and FROM address for emailing any errors
+email="emailsed"
+from="fromsed"
+
+x=""
+
+systemctl disable jackett --now
 sleep 5
 
 mv /opt/jackett /opt/jackettold
 
-( cd /tmp || return
-curl -s https://api.github.com/repos/Jackett/Jackett/releases | grep "browser_download_url".*Jackett.Binaries.LinuxAMDx64.tar.gz | head -1 | cut -d : -f 2,3 | tr -d \" | wget -i-
-tar -zxf Jackett.Binaries.LinuxAMDx64.tar.gz -C /opt/ )
-mv /opt/Jackett /opt/jackett
-chown -R jackett:jackett /opt/jackett
+while [[ ! -d /opt/Jackett && $x -lt 5 ]]; do
+  ( cd /tmp || exit 1
+  curl -s https://api.github.com/repos/Jackett/Jackett/releases | \
+  grep "browser_download_url.*Jackett.Binaries.LinuxAMDx64.tar.gz" | head -1 | cut -d : -f 2,3 | tr -d \" | wget -i- )
+  tar -zxf Jackett.Binaries.LinuxAMDx64.tar.gz -C /opt/
+  x=$(( x + 1 ))
+done
 
-rm -rf /opt/jackettold
+if [[ -d /opt/Jackett ]]; then
+  mv /opt/Jackett /opt/jackett
+  chown -R jackett:jackett /opt/jackett
+fi
 
-systemctl start jackett
+if [[ -d /opt/jackett ]]; then
+  rm -rf /opt/jackettold
+  rm /tmp/Jackett.Binaries.LinuxAMDx64.tar.gz
+  systemctl enable jackett --now
+else
+  echo "Jeckett update failed, reverting to previous version." | \
+  mail -s "Jackett update failed" \
+       -r "$from" \
+       "$email"
+  mv /opt/jackettold /opt/jackett
+  exit 1
+fi
